Medical Rag Chatbot/requirements.txt
langchain
langchain-community
langchain-huggingface
faiss-cpu
python-dotenv
pypdf
flask
sentence-transformers